LayoutTests/platform/wpe/storage/domstorage/localstorage/string-conversion-expected.txt

CONSOLE MESSAGE: QuotaExceededError: The quota has been exceeded.
This test case verifies that local storage only stores strings.
Length is 0
Testing implicit setters
